Tidy up sync source.
