+++ docusaurus.config.ts
@@ -1,11 +1,11 @@
 import {themes as prismThemes} from 'prism-react-renderer';
 import type {Config, LoadContext, PluginOptions} from '@docusaurus/types';
 import type * as Preset from '@docusaurus/preset-classic';
-import {loadSiteConfig} from "./src/builder/site-loader";
+import {loadSiteConfig} from "./src/builder/site-config-loader";
 import {buildScripts} from "./src/builder/scripts-builder";
 import {SCRIPTS_ROOT} from "./config/builder-config";
 import * as osPath from "path";
-import { Logger } from './src/builder/logger';
+import { Logger } from './src/builder/util/logger';
 
 const siteConfig = loadSiteConfig();
 Logger.instance.info(`🔧 Building site '${siteConfig.siteId}'`);
